+++ weekend-kotlin/src/main/java/com/github/aly8246/kotlin/pojo/User.java
@@ -1,0 +1,11 @@
+package com.github.aly8246.kotlin.pojo;
+
+import lombok.ToString;
+
+@ToString
+public class User {
+    private String id;
+    private String name;
+    private Integer userMoney;
+    private Integer age;
+}
